requirements.txt
Django>=5.0
dnspython
pymongo
djangorestframework
jdatetime
celery
redis
django-redis